pyproject.toml
[project]
name = "stock-agent"
version = "0.1.0"
description = "PoC: Karpathy-style LLM Wiki for Korean stock agent (NH internal)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "openai>=1.54",
  "python-dotenv>=1.0",
  "pydantic>=2.9",
  "jinja2>=3.1",
  "rank-bm25>=0.2.2",
  "numpy>=1.26",
  "python-frontmatter>=1.1",
  "rapidfuzz>=3.10",
  "diskcache>=5.6",
  "markdown>=3.5",
  "slowapi>=0.1.9",
  "certifi>=2024.2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
